Add table title with current date range to payment history table and cleaned up spacing
